+++ library/src/main/java/com/getstream/sdk/chat/rest/response/CompletableResponse.java
@@ -7,4 +7,8 @@ public class CompletableResponse {
     @SerializedName("duration")
     @Expose
     private String duration;
+
+    public String getDuration() {
+        return duration;
+    }
 }