+++ library/src/main/java/com/getstream/sdk/chat/view/MessageInputView.java
@@ -51,7 +51,9 @@ import com.getstream.sdk.chat.viewmodel.ChannelViewModel;
 import net.yslibrary.android.keyboardvisibilityevent.KeyboardVisibilityEvent;
 
 import java.io.File;
+import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.List;
 
 
 /**
@@ -489,9 +491,21 @@ public class MessageInputView extends RelativeLayout
         binding.ivSend.setEnabled(false);
 
         if (isEdit) {
-            getEditMessage().setText(input);
-            getEditMessage().setAttachments(messageInputClient.getSelectedAttachments());
-            viewModel.getChannel().updateMessage(getEditMessage(),  new MessageCallback() {
+            Message message = getEditMessage();
+            message.setText(input);
+            List<Attachment>newAttachments = messageInputClient.getSelectedAttachments();
+            if (newAttachments != null
+                    && !newAttachments.isEmpty()){
+                List<Attachment>attachments = message.getAttachments();
+                for (Attachment attachment : newAttachments){
+                    if (attachments == null)
+                        attachments = new ArrayList<>();
+                    attachments.add(attachment);
+                }
+                message.setAttachments(attachments);
+            }
+
+            viewModel.getChannel().updateMessage(message,  new MessageCallback() {
                 @Override
                 public void onSuccess(MessageResponse response) {
                     initSendMessage();