+++ library/src/main/java/com/getstream/sdk/chat/viewmodel/ChannelViewModel.java
@@ -41,7 +41,7 @@ import com.getstream.sdk.chat.rest.response.CompletableResponse;
 import com.getstream.sdk.chat.rest.response.EventResponse;
 import com.getstream.sdk.chat.rest.response.GetRepliesResponse;
 import com.getstream.sdk.chat.rest.response.MessageResponse;
-import com.getstream.sdk.chat.storage.Storage;
+import com.getstream.sdk.chat.storage.OnQueryListener;
 import com.getstream.sdk.chat.storage.Sync;
 import com.getstream.sdk.chat.utils.Constant;
 import com.getstream.sdk.chat.utils.MessageListItemLiveData;
@@ -74,7 +74,7 @@ public class ChannelViewModel extends AndroidViewModel implements LifecycleHandl
         this.channel = channel;
 
         // fetch offline messages
-        client().storage().selectChannelState(channel.getCid(), new Storage.OnQueryListener<ChannelState>() {
+        client().getStorage().selectChannelState(channel.getCid(), new OnQueryListener<ChannelState>() {
             @Override
             public void onSuccess(ChannelState channelState) {
                 Log.i(TAG, "Read messages from local cache...");
@@ -384,7 +384,7 @@ public class ChannelViewModel extends AndroidViewModel implements LifecycleHandl
     public void markLastMessageRead() {
         // this prevents infinite loops with mark read commands
         Message message = this.channel.getChannelState().getLastMessage();
-        if (message == null || !isEnableMarkRead()) {
+        if (message == null || !isEnableMarkRead() || message.getUserId().equals(client().getUserId())) {
             return;
         }
         if (lastMarkRead == null || message.getCreatedAt().getTime() > lastMarkRead.getTime()) {
@@ -978,16 +978,12 @@ public class ChannelViewModel extends AndroidViewModel implements LifecycleHandl
 
         if (message.getSyncStatus() == Sync.IN_MEMORY) {
             // insert the message into local storage
-            client().storage().insertMessageForChannel(channel, message);
+            client().getStorage().insertMessageForChannel(channel, message);
 
             // add the message here
             addMessage(message);
         }
 
-        // TODO: I'm reading this code and i don't get what it does, needs some clarification (Thierry)
-        if (client().isConnected()) {
-            channel.getChannelState().setReadDateOfChannelLastMessage(client().getUser(), message.getCreatedAt());
-        }
         // afterwards send the request
         channel.sendMessage(message,
                 new MessageCallback() {