+++ library/src/test/java/com/getstream/sdk/chat/rest/utils/TestCachedTokenProvider.java
@@ -1,0 +1,21 @@
+package com.getstream.sdk.chat.rest.utils;
+
+import com.getstream.sdk.chat.interfaces.CachedTokenProvider;
+import com.getstream.sdk.chat.interfaces.TokenProvider;
+
+/*
+ * Created by Anton Bevza on 2019-10-18.
+ */
+public class TestCachedTokenProvider implements CachedTokenProvider {
+    public static String TEST_TOKEN = "testToken";
+
+    @Override
+    public void getToken(TokenProvider.TokenProviderListener listener) {
+        listener.onSuccess(TEST_TOKEN);
+    }
+
+    @Override
+    public void tokenExpired() {
+        //ignore
+    }
+}